feat: 弹幕捕捉并输出

+++ src/main/java/config/BasicConfig.java
@@ -16,5 +16,5 @@ public class BasicConfig {
 
     private String accessSecret;
 
-    private int appId;
+    private Long appId;
 }
